pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "langchain-cerebras>=0.8.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
